requirements.txt
djangorestframework==3.14.0
oauthlib==3.2.0
